LOTY whitelist
LOTY is the official token of Polkalottery, the first cross-chain lottery on Polkadot. The token will be distributed via airdrop on Hydration on the same day as the TGE (with a 10% weekly vesting schedule). To ensure this process runs smoothly, we require a whitelist so users won’t need to activate degen mode. Currently, we have over 100 holders eligible for the initial airdrop, along with strong expectations from other users looking to buy (and sell) the token at the TGE